+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * fix(Range): add `aria-label` to thumb
 * test(FileUpload): add `aria-label` to axe test case
 * fix(ProseCard): change hover text color
+* docs(transformMDC): fix external links for self docs
 
 ## [2.1.6](https://github.com/bitrix24/b24ui/compare/v2.1.5...v2.1.6) (2025-11-20)